fix: address review comments for data formats and loaders
- WebDataset: cast Math.Min result to int for array allocation
- DatasetDownloader: throw on unexpected end of stream in SkipBytes and CopyBytes
- TransformedDataLoader: use ArrayPool for per-sample buffer to reduce allocations

+++ src/Data/Formats/WebDataset.cs
@@ -265,7 +265,7 @@ internal class WebDataset : IDisposable
             {
                 // Skip unknown entry types
                 long paddedSize = ((size + 511) / 512) * 512;
-                byte[] skip = new byte[Math.Min(4096, paddedSize)];
+                byte[] skip = new byte[(int)Math.Min(4096L, paddedSize)];
                 long remaining = paddedSize;
                 while (remaining > 0)
                 {
